% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{binarization}
\alias{binarization}
\title{Function to binarize vectors}
\usage{
binarization(vec)
}
\arguments{
\item{vec}{A vector of integers}
}
\description{
This function creates a one-hot encoded matrix given a vector.
}
\examples{
vector1 <- as.integer((rnorm(1000, 0, 1)  \%\% 1) * 10);
targets <- binarize(vector1);
}
\keyword{Binarization}
\keyword{One-Hot-Encode}

