% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{check_layers}
\alias{check_layers}
\title{Function to check the basic shape of data between Layers}
\usage{
check_layers(layers, dataset, target, batch_size)
}
\arguments{
\item{layers}{A list of descriptors}

\item{dataset}{A matrix with data, one example per row.}

\item{target}{A matrix with data targets, one example per row.}

\item{batch_size}{The batch_size to be used in the CNN or MLP}
}
\value{
TRUE if the list input and output dimensions match. Otherwise returns
FALSE.
}
\description{
This function checks if a list of layer descriptors is correctly dimensioned.
}
\examples{
dataset <- array(1:784000, c(1000, 1, 28, 28));
targets <- binarize(as.integer((rnorm(1000, 0, 1)  \%\% 1) * 10));
layers <- list(
             c("CONV", 1, 4, filter_size, 0.1, border_mode, batch_size),
             c("POOL", 4, 0.1, batch_size, win_size, stride),
             c("RELU", 4, batch_size),
             c("CONV", 4, 16, filter_size, 0.1, border_mode, batch_size),
             c("POOL", 16, 0.1, batch_size, win_size, stride),
             c("RELU", 16, batch_size),
             c("FLAT", 16, batch_size),
             c("LINE", 784, 64, 0.1, batch_size),
             c("RELV", batch_size),
             c("LINE", 64, 10, 0.1, batch_size),
             c("SOFT", 10, batch_size)
);
batch_size <- 10;
is_valid <- check_layers(layers, dataset, targets, batch_size);
}
\keyword{MLP}
