% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{conf.matrix}
\alias{conf.matrix}
\title{Function to produce a confusion matrix}
\usage{
conf.matrix(a, b)
}
\arguments{
\item{a}{First matrix to compare}

\item{b}{Second matrix to be compared}
}
\description{
This function creates a confusion matrix from two vectors/matrices with
output values. If the matrices are not "one-hot encoded", it takes the
maximum value per row as example output. The first matrix value is
represented as "rows" while the second matrix is represented as "columns".
}
\examples{
vector1 <- as.integer((rnorm(1000, 0, 1)  \%\% 1) * 10);
targets <- binarize(vector1);
}
\keyword{Confusion-Matrix}

