% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.cnn}
\alias{predict.cnn}
\title{Predicting using a Convolutional Network Function}
\usage{
\method{predict}{cnn}(cnn, newdata)
}
\arguments{
\item{cnn}{A trained CNN.}

\item{newdata}{A matrix with data, one example per row.}
}
\description{
This function predicts a dataset using a trained CNN. Admits as parameters
the testing dataset, and a trained CNN. Returns a matrix of predicted
outputs.
}
\examples{
test_X <- t(array(c(1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0,
                     0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0),
                     c(6, 1, 2, 3)));
results <- predict.cnn(cnn1, test_X);
}
\keyword{CNN}
