% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.mlp}
\alias{predict.mlp}
\title{Predicting using a MultiLayer Perceptron Network Function}
\usage{
\method{predict}{mlp}(mlp, newdata)
}
\arguments{
\item{mlp}{A trained MLP.}

\item{newdata}{A matrix with data, one example per row.}
}
\description{
This function predicts a dataset using a trained MLP. Admits as parameters
the testing dataset, and a trained MLP. Returns a matrix of predicted
outputs.
}
\examples{
test_X <- t(array(c(1, 1, 1, 0, 0, 0,
                    1, 0, 1, 0, 0, 0,
                    1, 1, 1, 0, 0, 0,
                    0, 0, 1, 1, 1, 0,
                    0, 0, 1, 0, 1, 0,
                    0, 0, 1, 1, 1, 0),
                    c(6, 6)));
results <- predict.cnn(mlp1, test_X);
}
\keyword{MLP}
